feat: add ts eslint, rewrite managing commands, refactor code
