+++ package.json
@@ -6,10 +6,10 @@
   "type": "module",
   "scripts": {
     "start": "npm run build && nodemon dist/index.js",
-    "build": "babel index.js -d dist",
-    "dev": "nodemon index.js",
-    "task1": "nodemon task1.js",
-    "task2": "nodemon task2.js",
+    "build": "babel src/index.js -d dist",
+    "dev": "nodemon src/index.js",
+    "task1": "nodemon src/task1.js",
+    "task2": "nodemon src/task2.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -26,13 +26,23 @@
   },
   "homepage": "https://github.com/intellectualDarknet/node-course#readme",
   "dependencies": {
+    "@types/node": "^18.11.10",
     "csvtojson": "^2.0.10",
-    "nodemon": "^2.0.20"
+    "eslint-config-prettier": "^8.5.0",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.8.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.19.3",
     "@babel/core": "^7.20.5",
     "@babel/node": "^7.20.5",
-    "@babel/preset-env": "^7.20.2"
+    "@babel/preset-env": "^7.20.2",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "eslint": "^8.28.0",
+    "eslint-config-standard-with-typescript": "^23.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.5.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "typescript": "^4.9.3"
   }
 }
